BusConnection::RegisterSignalHandler(): support multiple signal handlers
