+++ src/BusObjectWrapper.cc
@@ -96,9 +96,6 @@ NAN_METHOD(BusObjectWrapper::Signal) {
 }
 
 BusObjectImpl::BusObjectImpl(const char* path):ajn::BusObject(path){
-  loop = uv_default_loop();
-  signalCallback.callback = NULL;
-  uv_async_init(loop, &signal_async, signal_callback);
 }
 
 QStatus BusObjectImpl::AddInter(ajn::InterfaceDescription* interface){
@@ -106,52 +103,4 @@ QStatus BusObjectImpl::AddInter(ajn::InterfaceDescription* interface){
 }
 
 BusObjectImpl::~BusObjectImpl(){
-  if(signalCallback.callback){
-    delete signalCallback.callback;
-  }
 }
-
-void BusObjectImpl::signal_callback(uv_async_t *handle, int status) {
-    CallbackHolder* holder = (CallbackHolder*) handle->data;
-
-    v8::Local<v8::Object> msg = v8::Object::New();
-    size_t msgIndex = 0;
-    const ajn::MsgArg* arg = (*holder->message)->GetArg(msgIndex);
-    while(arg != NULL){
-      msgArgToObject(arg, msgIndex, msg);
-      msgIndex++;
-      arg = (*holder->message)->GetArg(msgIndex);
-    }
-
-    v8::Local<v8::Object> sender = v8::Object::New();
-    sender->Set(NanNew<v8::String>("sender"), NanNew<v8::String>((*holder->message)->GetSender()));
-    sender->Set(NanNew<v8::String>("session_id"), NanNew<v8::Integer>((*holder->message)->GetSessionId()));
-    sender->Set(NanNew<v8::String>("timestamp"), NanNew<v8::Integer>((*holder->message)->GetTimeStamp()));
-    sender->Set(NanNew<v8::String>("member_name"), NanNew<v8::String>((*holder->message)->GetMemberName()));
-    sender->Set(NanNew<v8::String>("object_path"), NanNew<v8::String>((*holder->message)->GetObjectPath()));
-    sender->Set(NanNew<v8::String>("signature"), NanNew<v8::String>((*holder->message)->GetSignature()));
-
-    v8::Handle<v8::Value> argv[] = {
-      msg,
-      sender
-    };
-    holder->callback->Call(2, argv);
-
-    if(holder->message){
-      delete holder->message;
-      holder->message = NULL;
-    }
-}
-
-void BusObjectImpl::ReceiveSignal(const ajn::InterfaceDescription::Member *member, const char *srcPath, ajn::Message &message){
-  if(signalCallback.callback){
-    signal_async.data = (void*) &signalCallback;
-    signalCallback.message = new ajn::Message(message);
-    uv_async_send(&signal_async);
-  }
-}
-
-void BusObjectImpl::SetSignalCallback(NanCallback* callback){
-  signalCallback.callback = callback;
-}
-
